+++ src/lib/src/config.ts
@@ -22,6 +22,7 @@ export type SpectatorOptions<T = any, H = HostComponent> = TestModuleMetadata & 
   disableAnimations?: boolean;
   host?: Type<H>;
   entryComponents?: any[];
+  componentProviders?: any[];
   mocks?: Type<any>[];
 };
 
@@ -30,12 +31,13 @@ const defaultOptions: SpectatorOptions<any, any> = {
   shallow: false,
   host: HostComponent,
   entryComponents: [],
+  componentProviders: [],
   mocks: []
 };
 
 export function initialModule<T, C = HostComponent>(options: SpectatorOptions<T, C> | Type<T>, withHost = false) {
   const merged = Object.assign({}, defaultOptions, options);
-  let moduleMetadata: TestModuleMetadata & { entryComponents: any[] };
+  let moduleMetadata: TestModuleMetadata & Partial<SpectatorOptions<T, C>>;
   let component;
   let host;
 
@@ -47,6 +49,7 @@ export function initialModule<T, C = HostComponent>(options: SpectatorOptions<T,
       imports: [NoopAnimationsModule],
       schemas: [],
       providers: [],
+      componentProviders: [],
       entryComponents: []
     };
   } else {
@@ -58,6 +61,7 @@ export function initialModule<T, C = HostComponent>(options: SpectatorOptions<T,
       imports: [merged.disableAnimations ? NoopAnimationsModule : [], ...(merged.imports || [])],
       schemas: [merged.shallow ? NO_ERRORS_SCHEMA : []],
       providers: [...(merged.providers || [])],
+      componentProviders: [merged.componentProviders || []],
       entryComponents: [merged.entryComponents]
     };